+++ components/collector/src/source_collectors/file_source_collectors/robot_framework.py
@@ -25,16 +25,21 @@ class RobotFrameworkTests(RobotFrameworkBaseClass):
 
     async def _parse_source_responses(self, responses: SourceResponses) -> SourceMeasurement:
         count = 0
+        total = 0
         entities = []
         test_results = cast(List[str], self._parameter("test_result"))
+        all_test_results = self._data_model["sources"][self.source_type]["parameters"]["test_result"]["values"]
         for response in responses:
             tree = await parse_source_response_xml(response)
             stats = tree.findall("statistics/total/stat")[1]
-            for test_result in test_results:
-                count += int(stats.get(test_result, 0))
-                for test in tree.findall(f".//test/status[@status='{test_result.upper()}']/.."):
-                    entities.append(Entity(key=test.get("id", ""), name=test.get("name", ""), test_result=test_result))
-        return SourceMeasurement(value=str(count), entities=entities)
+            for test_result in all_test_results:
+                total += int(stats.get(test_result, 0))
+                if test_result in test_results:
+                    count += int(stats.get(test_result, 0))
+                    for test in tree.findall(f".//test/status[@status='{test_result.upper()}']/.."):
+                        entities.append(
+                            Entity(key=test.get("id", ""), name=test.get("name", ""), test_result=test_result))
+        return SourceMeasurement(value=str(count), total=str(total), entities=entities)
 
 
 class RobotFrameworkSourceUpToDateness(RobotFrameworkBaseClass, SourceUpToDatenessCollector):