+++ components/collector/src/source_collectors/file_source_collectors/testng.py
@@ -17,11 +17,13 @@ class TestNGTests(XMLFileSourceCollector):
     async def _parse_source_responses(self, responses: SourceResponses) -> SourceMeasurement:
         test_statuses_to_count = cast(List[str], self._parameter("test_result"))
         test_count = 0
+        total = 0
         for response in responses:
             tree = await parse_source_response_xml(response)
             for test_status in test_statuses_to_count:
                 test_count += int(tree.get(test_status) or "0")
-        return SourceMeasurement(value=str(test_count))
+            total += int(tree.get("total") or "0")
+        return SourceMeasurement(value=str(test_count), total=str(total))
 
 
 class TestNGSourceUpToDateness(XMLFileSourceCollector, SourceUpToDatenessCollector):